add key exists check

+++ scripts/install-registry-tweaks.ps1
@@ -25,7 +25,9 @@ New-ItemProperty -Path $DynamicSearch_Path -Name $DynamicSearch_Name -Value 0 -T
 
 # Delete the recycle bin desktop shortcut
 $RecycleBin_Path = 'HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\Desktop\NameSpace\{645FF040-5081-101B-9F08-00AA002F954E}'
-Remove-Item -Path $RecycleBin_Path -Recurse -Force
+If (Test-Path $RecycleBin_Path) {
+  Remove-Item -Path $RecycleBin_Path -Recurse -Force
+}
 
 $ChatIcon_Path = 'HKLM:\SOFTWARE\Policies\Microsoft\Windows\Windows Chat'
 $ChatIcon_Name = 'ChatIcon'
